+++ CleverPush/Source/CleverPushSelectorHelpers.m
@@ -3,76 +3,29 @@
 
 #import "CleverPushSelectorHelpers.h"
 
-#pragma mark - Instances Overrises selector.
-BOOL checkIfInstanceOverridesSelector(Class instance, SEL selector) {
-    Class instSuperClass = [instance superclass];
-    return [instance instanceMethodForSelector: selector] != [instSuperClass instanceMethodForSelector: selector];
-}
-
-#pragma mark - get Class With Protocol In Hierarchy
-Class getClassWithProtocolInHierarchy(Class searchClass, Protocol* protocolToFind) {
-    if (!class_conformsToProtocol(searchClass, protocolToFind)) {
-        if ([searchClass superclass] == nil)
-            return nil;
-        Class foundClass = getClassWithProtocolInHierarchy([searchClass superclass], protocolToFind);
-        if (foundClass)
-            return foundClass;
-        return searchClass;
-    }
-    return searchClass;
-}
-
 #pragma mark - inject Selector
-BOOL injectSelector(Class newClass, SEL newSel, Class addToClass, SEL makeLikeSel) {
-    Method newMeth = class_getInstanceMethod(newClass, newSel);
+BOOL injectSelector(Class targetClass, SEL targetSelector, Class myClass, SEL mySelector) {
+    Method newMeth = class_getInstanceMethod(myClass, mySelector);
     IMP imp = method_getImplementation(newMeth);
-    
+
     const char* methodTypeEncoding = method_getTypeEncoding(newMeth);
-    
-    BOOL existing = class_getInstanceMethod(addToClass, makeLikeSel) != NULL;
-    
+    BOOL existing = class_getInstanceMethod(targetClass, targetSelector) != NULL;
+
     if (existing) {
-        class_addMethod(addToClass, newSel, imp, methodTypeEncoding);
-        newMeth = class_getInstanceMethod(addToClass, newSel);
-        Method orgMeth = class_getInstanceMethod(addToClass, makeLikeSel);
+        Method orgMeth = class_getInstanceMethod(targetClass, targetSelector);
+        IMP orgImp = method_getImplementation(orgMeth);
+
+        if (imp == orgImp) {
+            return existing;
+        }
+
+        class_addMethod(targetClass, mySelector, imp, methodTypeEncoding);
+        newMeth = class_getInstanceMethod(targetClass, mySelector);
         method_exchangeImplementations(orgMeth, newMeth);
     }
-    else
-        class_addMethod(addToClass, makeLikeSel, imp, methodTypeEncoding);
-    
+    else {
+        class_addMethod(targetClass, targetSelector, imp, methodTypeEncoding);
+    }
+
     return existing;
-}
-
-void injectToProperClass(SEL newSel, SEL makeLikeSel, NSArray* delegateSubclasses, Class myClass, Class delegateClass) {
-    for(Class subclass in delegateSubclasses) {
-        if (checkIfInstanceOverridesSelector(subclass, makeLikeSel)) {
-            injectSelector(myClass, newSel, subclass, makeLikeSel);
-            return;
-        }
-    }
-    injectSelector(myClass, newSel, delegateClass, makeLikeSel);
-}
-
-NSArray* ClassGetSubclasses(Class parentClass) {
-    int numClasses = objc_getClassList(NULL, 0);
-    Class *classes = (Class*)malloc(sizeof(Class) * numClasses);
-    
-    objc_getClassList(classes, numClasses);
-    
-    NSMutableArray *result = [NSMutableArray array];
-    
-    for (NSInteger i = 0; i < numClasses; i++) {
-        Class superClass = classes[i];
-        
-        while(superClass && superClass != parentClass) {
-            superClass = class_getSuperclass(superClass);
-        }
-        
-        if (superClass)
-            [result addObject:classes[i]];
-    }
-    
-    free(classes);
-    
-    return result;
 }